refactor: update package.json scripts, enhance Vite config for bundle analysis, and implement lazy loading for Viewer component
