+++ packages/ui/src/components/Viewer/labelStyle.ts
@@ -1,0 +1,14 @@
+import { Vector2 } from "3dmol";
+
+export const labelStyle = {
+  alignment: "center",
+  screenOffset: new Vector2(28, -28),
+  backgroundColor: "#fff",
+  backgroundOpacity: 0.9,
+  borderColor: "#999",
+  borderThickness: 1.5,
+  font: "'Inter', sans-serif",
+  fontColor: "#000",
+  fontSize: 12.6,
+  inFront: true,
+};
