+++ packages/ui/src/index.tsx
@@ -75,7 +75,7 @@ export { default as SummaryStatsTable } from "./components/SummaryStatsTable";
 export * from "./components/Table";
 export { default as Tooltip } from "./components/Tooltip";
 export { default as TooltipStyledLabel } from "./components/TooltipStyledLabel";
-export { default as Viewer } from "./components/Viewer/Viewer";
+export { default as Viewer } from "./components/Viewer/LazyViewer";
 export { default as ViewerDropdown } from "./components/Viewer/ViewerDropdown";
 export { default as ViewerRadios } from "./components/Viewer/ViewerRadios";
 export { default as ViewerTrack } from "./components/Viewer/ViewerTrack";
